Move LCU to Core package
