+++ src-tauri/src/core.rs
@@ -1,2 +1,3 @@
 pub mod data_dragon;
 pub mod helpers;
+pub mod lcu;
