+++ src-tauri/src/extensions.rs
@@ -1,2 +1,1 @@
 pub mod ugg;
-pub mod lcu;
